Moving the fake buffer pool implementation to it's correct package
